+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ## [Unreleased]
 
 
+## [v317] - 2025-08-07
+
+- JRuby 10.0.2.0 is now available
+
+
+## [v316] - 2025-08-06
+
+- Explicitly error on end-of-life (EOL) stack `heroku-20` (https://github.com/heroku/heroku-buildpack-ruby/pull/1629)
+- Default Ruby version is now 3.3.9 (https://github.com/heroku/heroku-buildpack-ruby/pull/1624)
+
 ## [v315] - 2025-07-24
 
 - Ruby 3.2.9 and 3.3.9 is now available
@@ -1735,7 +1745,9 @@ Bugfixes:
 * Change gem detection to use lockfile parser
 * use `$RACK_ENV` when thin is detected for rack apps
 
-[unreleased]: https://github.com/heroku/heroku-buildpack-ruby/compare/v315...main
+[unreleased]: https://github.com/heroku/heroku-buildpack-ruby/compare/v317...main
+[v317]: https://github.com/heroku/heroku-buildpack-ruby/compare/v316...v317
+[v316]: https://github.com/heroku/heroku-buildpack-ruby/compare/v315...v316
 [v315]: https://github.com/heroku/heroku-buildpack-ruby/compare/v314...v315
 [v314]: https://github.com/heroku/heroku-buildpack-ruby/compare/v313...v314
 [v313]: https://github.com/heroku/heroku-buildpack-ruby/compare/v312...v313